greenDao集成使用

+++ app/src/main/java/com/fengkai/zhouyang/yangyanghongkong/addprodut/port/IProductModel.java
@@ -9,5 +9,5 @@ public interface IProductModel {
 
     List<Product> queryAllData();
 
-    void deleteProductById(int id);
+    void deleteProductById(long id);
 }
